post creator working
